requirements.txt
Werkzeug
Flask
numpy
Keras
gevent
pillow
h5py
tensorflow
pandas
lxml.html
requests